added pars$tina$do.clustering
This is a switch (true/false) to turn clustering on/off. It only affects tina() explicitly, where the clustering step is skipped, a trivial clustering result is returned (all features; feature in each cluster) and the clustering method is recorded as 'none'. Matching (@setup) is agnostic to whether the clustering is trivial or true. The clustering method is used later in filter_matches() to determine whether propagate_matches() runs - after that, clusters are no longer relevant.

By not clustering, compute times for matching will increase (scaling with number of features) and much of this will be wasteful because of feature duplication, but the quality of the matches will likely be a little better as each is tailored to the details of each feature. Additionally, there is less chance for introduction of artifacts and the procedure is less complicated.