score_matches not running on 424
Backfit calculation erroring and returning NULL for entire core on several of these. This could happen if the fit vals for a spectrum are NULL, or if bff resonance extraction fails.

Put in several checks to catch these instances and return empty (NA) rows, which are then filtered out before concatenating to chunk. NOTE: the effective subset will change as a result of failed fits, as those spectra will be excluded for that ref feature.

additionally, encountered error in showing evidence were lib.info was accessed using the absolute filepath provided in params.yaml. This won't be portable, and lib.info is copied into the wd at spinup. So now, using relative filepath.