Update filter.matches_shiftDelta.R
adding the ref to the region WILL SHIFT everything with a large ref number!!!
definite bug, and this will affect a lot of results because of filtering. Things with bigger ref numbers (at the end of the library list) will be affected most, and retained matches will be shifted to the right. this explains a lot!