trying to figure out filtering (singlets)
singlet-like things not getting filtered out

simple peaks (unsurprisingly) not enough to determine singleticity.

what if we remove each simple peak and assess its effect on essence of the feature remaining? erode iteratively removes all simple peaks by interpolating between valleys. If strong bl effect, then this has no effect for a while. If not, the feature degrades rapidly.
erode2 iteratively removes the smallest (prominence) peak, progressively smoothing/linearizing the feature. when a region is removed, it becomes clearer what its effect was on the shape. If removal of a peak results in a huge drop in variance of the feature described, that region should not be replaced by a linear interpolation. however, replacements with minimal effect should be replaced. at the end of this, we can assess how many "important" peaks there are"